requirements.txt
certifi==2024.2.2
charset-normalizer==3.3.2
decorator==4.4.2
ffmpeg==1.4
filelock==3.14.0
fsspec==2024.3.1
idna==3.7
imageio==2.34.1
imageio-ffmpeg==0.4.9
Jinja2==3.1.3
llvmlite==0.42.0
MarkupSafe==2.1.5
more-itertools==10.2.0
moviepy==1.0.3
mpmath==1.3.0
networkx==3.3
numba==0.59.1
numpy==1.26.4
openai-whisper @ git+https://github.com/openai/whisper.git@ba3f3cd54b0e5b8ce1ab3de13e32122d0d5f98ab
pillow==10.3.0
proglog==0.1.10
pydub==0.25.1
regex==2024.4.28
requests==2.31.0
setuptools==69.5.1
SpeechRecognition==3.10.3
sympy==1.12
tiktoken==0.6.0
torch==2.3.0
tqdm==4.66.2
typing_extensions==4.11.0
urllib3==2.2.1
